[0908_Make load_fIle Function and Delete the String Copy Work in insert_line Function] 1. I made the new function, load_file to make the view_file be more independent. The previous code let the view_file function save the text file into the memory and display them on the prompt at the same time. Today I seperated the two steps by making the load_file function to save the whole text file into the memory and to call the print_line funciton during the view_file running time.
2. I omitted the string copy work in insert_line function. After I printed the character '`' to indicate the line change,
there was a problem that the character ('`') is double and triple printed every time I open the file after insertion and append work. So I deleted the course to insert the character('`') instead of line-feed character. Then I ordered the print_line function to print it('`') if there is a line-feed character.
In conclusion, I don't put the character('`') into the memory directly but print it when it meets the condition that there was a line change.